UseSharedResourcesMain - Design consistent .NET MAUI XAML pages by using shared resouces
Use ResourceDictionary and DynamicResource to change Light and Dark color.

It's similar to StaticResource in that it does a dictionary lookup when the target object is created. But it also listens for changes to the resource in the dictionary. If the resource value in the dictionary changes, DynamicResource automatically updates the UI.